+++ tools/serve.js
@@ -1,0 +1,8 @@
+const esbuild = require('esbuild');
+const { config } = require('./config');
+
+esbuild.build({
+  ...config,
+  watch: true,
+  sourcemap: true,
+});